moved default parameters to config
