+++ src/abatement/abatement_enerdata.py
@@ -13,7 +13,7 @@ from typing import Any
 from scipy.interpolate import interp1d
 import numpy as np
 
-from src.default_parameters import AbatementDefaults
+from config.default_parameters import AbatementDefaults
 from src.util.enumerations import Abatement
